fix: gamecontroller processes

+++ controllers/gamecontroller.js
@@ -1,10 +1,12 @@
 var router = require('express').Router();
-var Game = require('../models/game');
+var sequelize = require('../db');
+const Sequelize = require('sequelize');
+var Game = require('../models/game')(sequelize, Sequelize.DataTypes);
 
 router.get('/all', (req, res) => {
     Game.findAll({ where: { owner_id: req.user.id } })
         .then(
-            function findSuccess(data) {
+            function findSuccess(games) {
                 res.status(200).json({
                     games: games,
                     message: "Data fetched."
@@ -23,12 +25,19 @@ router.get('/:id', (req, res) => {
     Game.findOne({ where: { id: req.params.id, owner_id: req.user.id } })
         .then(
             function findSuccess(game) {
-                res.status(200).json({
-                    game: game
-                })
+                if (game) {
+                    res.status(200).json({
+                        game: game
+                    })
+                }
+                else {
+                    res.status(404).json({
+                        message: "Game not found."
+                    })
+                }
             },
 
-            function findFail(err) {
+            function findFail() {
                 res.status(500).json({
                     message: "Data not found."
                 })
@@ -39,7 +48,7 @@ router.get('/:id', (req, res) => {
 router.post('/create', (req, res) => {
     Game.create({
         title: req.body.game.title,
-        owner_id: req.body.user.id,
+        owner_id: req.user.id,
         studio: req.body.game.studio,
         esrb_rating: req.body.game.esrb_rating,
         user_rating: req.body.game.user_rating,
@@ -68,17 +77,26 @@ router.put('/update/:id', (req, res) => {
         have_played: req.body.game.have_played
     },
         {
+            returning: true,
+            plain: true,
             where: {
                 id: req.params.id,
-                owner_id: req.user
+                owner_id: req.user.id
             }
         })
         .then(
-            function updateSuccess(game) {
-                res.status(200).json({
-                    game: game,
-                    message: "Successfully updated."
-                })
+            function updateSuccess(updatedGame) {
+                if (updatedGame) {
+                    res.status(200).json({
+                        updatedGame: updatedGame[1],
+                        message: "Successfully updated."
+                    })
+                }
+                else {
+                    res.status(404).json({
+                        message: "Game not found."
+                    })
+                }
             },
 
             function updateFail(err) {
@@ -91,26 +109,47 @@ router.put('/update/:id', (req, res) => {
 })
 
 router.delete('/remove/:id', (req, res) => {
-    Game.destroy({
+    Game.findOne({
         where: {
-            id: req.params.id,
-            owner_id: req.user.id
+            id: req.params.id, owner_id: req.user.id
         }
     })
-    .then(
-        function deleteSuccess(game) {
-            res.status(200).json({
-                game: game,
-                message: "Successfully deleted"
-            })
-        },
+        .then(
+            function findSuccess(deletedGame) {
+                if (deletedGame) {
+                    Game.destroy({
+                        where: {
+                            id: req.params.id,
+                            owner_id: req.user.id
+                        }
+                    })
+                        .then(
+                            function deleteSuccess() {
+                                res.status(200).json({
+                                    deletedGame: deletedGame,
+                                    message: "Successfully deleted"
+                                })
+                            },
 
-        function deleteFail(err) {
-            res.status(500).json({
-                error: err.message
-            })
-        }
-    )
-})
+                            function deleteFail(err) {
+                                res.status(500).json({
+                                    error: err.message
+                                })
+                            }
+                        );
+                }
+                else {
+                    res.status(404).json({
+                        message: "Game not found."
+                    })
+                }
+            },
+            function findFail() {
+                res.status(500).json({
+                    message: "Data not found."
+                })
+            }
+        );
+});
 
 module.exports = router;
